fix AndroidLogger

+++ src/SDLApplication/AndroidLogger.cpp
@@ -15,12 +15,12 @@ namespace Mengine
     {
     }
     //////////////////////////////////////////////////////////////////////////
-    void AndroidLogger::log( ELoggerLevel _level, uint32_t _flag, const Char * _data, uint32_t _count )
+    void AndroidLogger::log( ELoggerLevel _level, uint32_t _flag, const Char * _data, size_t _count )
     {
 		MENGINE_UNUSED(_level);
         MENGINE_UNUSED(_flag);
 
-        __android_log_print( ANDROID_LOG_ERROR, "Mengine", "%.*s", _count, _data );
+        __android_log_print( ANDROID_LOG_ERROR, "Mengine", "%.*s", (int)_count, _data );
     }
     //////////////////////////////////////////////////////////////////////////
     void AndroidLogger::flush()
